Fixed signed/unsigned/size_t warnings in the Matlab toolbox

+++ Cantera/matlab/src/transportmethods.cpp
@@ -20,7 +20,7 @@ void reportError();
             char* model = getString(prhs[3]);
             int loglevel = getInt(prhs[4]);            
             int m = -2;
-            m = newTransport(model, n, loglevel);
+            m = (int) newTransport(model, n, loglevel);
             if (m < 0) reportError();
             
             // Create matrix for the return argument.
